change stronghold crate's structure, revert persist changes on drop
