+++ identity_stronghold/src/lib.rs
@@ -2,13 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) mod ed25519;
-mod stronghold_jwk_storage;
-#[cfg(any(feature = "bbs-plus", test))]
-mod stronghold_jwk_storage_bbs_plus_ext;
-mod stronghold_key_id;
+mod storage;
 pub(crate) mod stronghold_key_type;
 #[cfg(test)]
 mod tests;
 pub(crate) mod utils;
 
-pub use stronghold_jwk_storage::*;
+pub use storage::*;
