+++ identity_stronghold/src/storage/stronghold_key_id.rs
